requirements.txt
botasaurus
botasaurus_server
dateutils
lxml
regex

fastapi
uvicorn
bottle
python-dotenv